Add files via upload
Create a folder in file explorer to place contents. You will need to include your own API Key in the actual syntax to see the functionality of our app in its entirety. In the folder created for this program: (i) create a new folder called "frame0" which needs to include all .png files in this repo, then (ii) create a new folder called "_pycache_" which needs to include all .pyc files in this repo, (iii) .DS_STORE file can be its own instance under the folder created for this program.